change minBet to int

+++ casino/src/main/java/game/Blackjack.java
@@ -1,7 +1,7 @@
 package game;
 
 public class Blackjack extends GameAbstract {
-    public Blackjack(int maxPlayers, double minBet, double cashOutMult, double winChance, int gameTime){
+    public Blackjack(int maxPlayers, int minBet, double cashOutMult, double winChance, int gameTime){
         super("Blackjack", maxPlayers, minBet, cashOutMult, winChance, gameTime);
     }
